keep interface and kernel seperate
